refactor: move slash command declarations
